+++ src/main/java/ui/Main.java
@@ -1,0 +1,7 @@
+package ui;
+
+public class Main{
+	public static void main(String[] args) {
+		javafx.application.Application.launch(UI.class);
+	}
+}